\name{actionet-package}
\alias{actionet-package}
\alias{actionet}
\docType{package}
\title{
\packageTitle{actionet}
}
\description{
\packageDescription{actionet}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{actionet}

Maintainer: \packageMaintainer{actionet}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = actionet)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{actionet}}
%\section{Documentation Index}{\packageIndices{actionet}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
