% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_specificity.R
\name{archetypeFeatureSpecificity}
\alias{archetypeFeatureSpecificity}
\title{Computes feature specificity scores for each archetype.}
\usage{
archetypeFeatureSpecificity(
  adata = NULL,
  layer = "logcounts",
  map_slot = "archetype_footprint",
  map_out_prefix = "archetype",
  thread_no = 0,
  return_raw = FALSE,
  assay_name = NULL,
  ace = NULL
)
}
\description{
Computes feature specificity scores for each archetype.
}
