% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{archetypeFeatureSpecificitySparse}
\alias{archetypeFeatureSpecificitySparse}
\title{Compute feature specificity (from archetype footprints)}
\usage{
archetypeFeatureSpecificitySparse(S, H, thread_no = 0L)
}
\arguments{
\item{S}{Input matrix (sparseMatrix)}

\item{H}{A soft membership matrix - Typically H_merged from the mergeArchetypes() function.}
}
\value{
A list with the over/under-logPvals
}
\description{
Compute feature specificity (from archetype footprints)
}
\examples{
prune.out = collectArchetypes(ACTION.out$C, ACTION.out$H)
G = buildNetwork(prune.out$H_stacked)
unification.out = mergeArchetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
S.norm = renormalize_input_matrix(S, cell.clusters)
logPvals.list = compute_archetype_feature_specificity(S.norm, unification.out$H_merged)
specificity.scores = logPvals.list$upper_significance
}
