% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{assess.peakset.enrichment.from.archetypes}
\alias{assess.peakset.enrichment.from.archetypes}
\title{Performs geneset enrichment analysis on archetypes}
\usage{
assess.peakset.enrichment.from.archetypes(
  adata = NULL,
  associations,
  min.counts = 0,
  specificity.slot = "archetype_feat_specificity_upper",
  ace = NULL
)
}
\arguments{
\item{adata}{AnnData or compatible ACTIONet output object.}

\item{associations}{Either a genes x pathways membership matrix, or a set of genesets}

\item{L}{Maximum length of the top-ranked genes to consider}
}
\value{
Matrix pathway x cell type/states
}
\description{
Performs geneset enrichment analysis on archetypes
}
\examples{
data("gProfilerDB_human")
associations <- gProfilerDB_human$SYMBOL$WP
Geneset.enrichments <- assess.peakset.enrichment.from.archetypes(adata, associations)
}
