% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{assess.TF.activities.from.archetypes}
\alias{assess.TF.activities.from.archetypes}
\title{Compute the activity score of each TF based on the observed activity of its targets.
(Similar to ChEA3; it is based on a meta-analysis using 4 separate datasets)}
\usage{
assess.TF.activities.from.archetypes(adata = NULL, ace = NULL)
}
\arguments{
\item{adata}{AnnData or compatible ACTIONet output object.}
}
\value{
Matrix of TF x archetypes indicating inferred TF activity scores
}
\description{
Compute the activity score of each TF based on the observed activity of its targets.
(Similar to ChEA3; it is based on a meta-analysis using 4 separate datasets)
}
\examples{
TF.scores <- assess.TF.activities.from.archetypes(adata)
}
