% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_assess_enrichment}
\alias{C_assess_enrichment}
\title{Computes feature enrichment wrt a given annotation}
\usage{
C_assess_enrichment(scores, associations, thread_no = 0L)
}
\arguments{
\item{scores}{Specificity scores of features}

\item{associations}{Binary matrix of annotations}

\item{L}{Length of the top-ranked scores to scan}
}
\value{
Matrix of log-pvalues
}
\description{
Computes feature enrichment wrt a given annotation
}
\examples{
data("gProfilerDB_human")
G = colNets(ace)$ACTIONet
associations = gProfilerDB_human$SYMBOL$REAC
common.genes = intersect(rownames(ace), rownames(associations))
specificity_scores = rowFactors(ace)[["H_merged_upper_significance"]]
logPvals = computeFeatureSpecificity(
specificity_scores[common.genes, ], annotations[common.genes, ]
)
rownames(logPvals) = colnames(specificity_scores)
colnames(logPvals) = colnames(annotations)
}
