% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_buildNetwork}
\alias{C_buildNetwork}
\title{Builds an interaction network from the multi-level archetypal decompositions}
\usage{
C_buildNetwork(
  H,
  algorithm = "k*nn",
  distance_metric = "jsd",
  density = 1,
  thread_no = 0L,
  M = 16,
  ef_construction = 200,
  ef = 200,
  mutual_edges_only = TRUE,
  k = 10L
)
}
\arguments{
\item{density}{Overall density of constructed graph. The higher the density,
the more edges are retained (default = 1.0).}

\item{thread_no}{Number of parallel threads (default = 0).}

\item{mutual_edges_only}{Symmetrization strategy for nearest-neighbor edges.
If it is true, only mutual nearest-neighbors are returned (default=TRUE).}

\item{H_stacked}{Output of the collectArchetypes() function.}
}
\value{
G Adjacency matrix of the ACTIONet graph.
}
\description{
Builds an interaction network from the multi-level archetypal decompositions
}
\examples{
prune.out = collectArchetypes(ACTION.out$C, ACTION.out$H)
G = buildNetwork(prune.out$H_stacked)
}
