% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_computeGroupedRowMeansSparse}
\alias{C_computeGroupedRowMeansSparse}
\title{Average matrix within groups}
\usage{
C_computeGroupedRowMeansSparse(S, sample_assignments)
}
\arguments{
\item{S}{matrix}

\item{sample_assignments}{Vector of column groupings. Group labels must be continuous integers or coercible to such.}
}
\value{
S matrix with columns of values average within each group of sample_assignments
}
\description{
Average matrix within groups
}
