% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_computeGroupedRowSumsSparse}
\alias{C_computeGroupedRowSumsSparse}
\title{Aggregate matrix within groups}
\usage{
C_computeGroupedRowSumsSparse(S, sample_assignments)
}
\arguments{
\item{S}{matrix of type "dMatrix"}

\item{sample_assignments}{Vector of column groupings. Group labels must be continuous integers or coercible to such.}
}
\value{
S matrix with columns of values aggregated within each group of sample_assignments
}
\description{
Aggregate matrix within groups
}
