% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_computeGroupedSumsSparse}
\alias{C_computeGroupedSumsSparse}
\title{Aggregate matrix within groups (rowwise or columnwise)}
\usage{
C_computeGroupedSumsSparse(S, sample_assignments, axis = 0L)
}
\arguments{
\item{S}{matrix of type "dMatrix"}

\item{sample_assignments}{Vector of groupings. Group labels must be continuous integers or coercible to such.}

\item{axis}{0 for rowwise (default), 1 for columnwise}
}
\value{
Aggregated matrix
}
\description{
Aggregate matrix within groups (rowwise or columnwise)
}
