% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_computeGroupedVarsSparse}
\alias{C_computeGroupedVarsSparse}
\title{Variance matrix within groups (rowwise or columnwise)}
\usage{
C_computeGroupedVarsSparse(S, sample_assignments, axis = 0L)
}
\arguments{
\item{S}{matrix}

\item{sample_assignments}{Vector of groupings. Group labels must be continuous integers or coercible to such.}

\item{axis}{0 for rowwise (default), 1 for columnwise}
}
\value{
Variance matrix
}
\description{
Variance matrix within groups (rowwise or columnwise)
}
