% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_computeNetworkDiffusion}
\alias{C_computeNetworkDiffusion}
\title{Computes network diffusion over a given network, starting with an arbitrarty
set of initial scores}
\usage{
C_computeNetworkDiffusion(
  G,
  X0,
  alpha = 0.85,
  max_it = 5L,
  thread_no = 0L,
  approx = FALSE,
  norm_method = 0L,
  tol = 1e-08
)
}
\arguments{
\item{G}{Input graph}

\item{X0}{Matrix of initial values per diffusion (ncol(G) == nrow(G) == ncol(X0))}

\item{alpha}{Random-walk depth ( between [0, 1] ) ' @param max_it PageRank iterations}

\item{thread_no}{Number of parallel threads (default=0)}
}
\value{
Matrix of diffusion scores
}
\description{
Computes network diffusion over a given network, starting with an arbitrarty
set of initial scores
}
\examples{
G = colNets(ace)$ACTIONet
gene.expression = Matrix::t(logcounts(ace))[c("CD19", "CD14", "CD16"), ]
smoothed.expression = computeNetworkDiffusionApprox(G, gene.expression)
}
