% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_fastSpMatViewSum}
\alias{C_fastSpMatViewSum}
\title{Fast column/row sums of sparse submatrix view}
\usage{
C_fastSpMatViewSum(X, idx, dim = 2L, threads = 1L)
}
\arguments{
\item{X}{Armadillo sparse matrix}

\item{idx}{Vector of row or column indices (1-index)}

\item{dim}{Dimensions (R conventon) along which to compute sums (1: `rowSums()`, 2: `colSums()`;)}

\item{threads}{Threads to use for parallel subsetting and summation. Quickly depreciating (<=4 recommended)}
}
\value{
R vector or column/row sums.
}
\description{
Fast column/row sums of sparse submatrix view
}
