% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_mergeArchetypes}
\alias{C_mergeArchetypes}
\title{Identify and merge redundant archetypes into a representative subset}
\usage{
C_mergeArchetypes(S_r, C_stacked, H_stacked, thread_no = 0L)
}
\arguments{
\item{S_r}{Reduced data matrix from which archetypes were found.}

\item{C_stacked}{Concatenated (and filtered) <code>C</code> (<b>S_r.n</b> x <em>n</em>) matrix.
Output of <code>collectArchetypes()</code> in <code>ResCollectArch["C_stacked"]</code>.}

\item{H_stacked}{Concatenated (and filtered) <code>H</code> (<b>S_r.n</b> x <em>n</em>) matrix.
Output of <code>collectArchetypes()</code> in <code>ResCollectArch["H_stacked"]</code>.}

\item{thread_no}{Number of CPU threads to use. If 0, number is automatically determined.}

\item{normalization}{Normalization method to apply to <b>S_r</b>.}
}
\value{
A named list: \itemize{
\item archetype_groups: Equivalent classes of archetypes (non-redundant)
\item C_merged,H_merged: C and H matrices of merged archetypes
\item sample_assignments: Assignment of samples/cells to merged archetypes
}
}
\description{
Identify and merge redundant archetypes into a representative subset
}
\examples{
prune.out = collectArchetypes(ACTION.out$C, ACTION.out$H)
G = buildNetwork(prune.out$H_stacked)
unification.out = mergeArchetypes(G, S_r, prune.out$C_stacked, prune.out$H_stacked)
cell.clusters = unification.out$sample_assignments
}
