% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_runAA}
\alias{C_runAA}
\title{Runs archetypal analysis (AA)}
\usage{
C_runAA(A, W0, max_it = 100L, tol = 1e-06)
}
\arguments{
\item{A}{Input matrix.}

\item{W0}{Matrix with <em>k</em> columns representing initial archetypes.}

\item{max_it}{Maximum number of iterations.}

\item{tol}{Convergence tolerance.}
}
\value{
Field with matrices <b>C</b> (<b>A.n_cols</b> x <em>k</em>) and <b>H</b> (<em>k</em> x <b>A.n_cols</b>).
}
\description{
Runs archetypal analysis (AA)
}
\examples{
S_r = t(reducedDims(ace)$ACTION)
SPA.out = runSPA(S_r, 10)
W0 = S_r[, SPA.selected_cols]
AA.out = runAA(S_r, W0)
H = AA.out$H
cell.assignments = apply(H, 2, which.max)
}
