% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{collect_archetypes}
\alias{collect_archetypes}
\title{Filter and aggregate multi-level archetypes}
\usage{
collect_archetypes(C_trace, H_trace, spec_th = -3, min_obs = 3L)
}
\arguments{
\item{C_trace}{Field containing C matrices. Output of <code>run_ACTION()</code> in <code>ResACTION["C"]</code>.}

\item{H_trace}{Field containing H matrices. Output of <code>run_ACTION()</code> in <code>ResACTION["H"]</code>.}

\item{spec_th}{Minimum threshold (as z-score) to filter archetypes by specificity.}

\item{min_obs}{Minimum number of observations assigned to an archetypes needed to retain that archetype.}
}
\value{
A named list: \itemize{
\item selected_archs: List of final archetypes that passed the
filtering/pruning step.
\item C_stacked,H_stacked: Horizontal/Vertical
concatenation of filtered C and H matrices, respectively.
}
}
\description{
Filter and aggregate multi-level archetypes
}
\examples{
S = logcounts(sce)
reduction.out = reduce(S, reduced_dim = 50)
S_r = reduction.out$S_r
ACTION.out = run_ACTION(S_r, k_max = 10)
reconstruction.out = reconstruct_archetypes(S, ACTION.out$C, ACTION.out$H)
}
