% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeArchetypeCentrality}
\alias{computeArchetypeCentrality}
\title{Compute coreness of subgraph vertices induced by each archetype}
\usage{
computeArchetypeCentrality(G, sample_assignments)
}
\arguments{
\item{G}{Input graph}

\item{sample_assignments}{Archetype discretization (output of mergeArchetypes())}
}
\value{
cn core-number of each graph node
}
\description{
Compute coreness of subgraph vertices induced by each archetype
}
\examples{
G = colNets(ace)$ACTIONet
assignments = ace$archetype.assignment
connectivity = computeCoreness(G, assignments)
}
