% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeCoreness}
\alias{computeCoreness}
\title{Compute coreness of graph vertices}
\usage{
computeCoreness(G)
}
\arguments{
\item{G}{Input graph}
}
\value{
cn core-number of each graph node
}
\description{
Compute coreness of graph vertices
}
\examples{
G = colNets(ace)$ACTIONet
cn = computeCoreness(G)
}
