% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_specificity.R
\name{computeFeatureSpecificity}
\alias{computeFeatureSpecificity}
\title{Computes feature specificity scores for each cluster.}
\usage{
computeFeatureSpecificity(
  adata = NULL,
  labels,
  labels_use = NULL,
  layer = "logcounts",
  map_out_prefix = "cluster",
  return_lower = FALSE,
  thread_no = 0,
  return_raw = FALSE,
  assay_name = NULL,
  obj = NULL
)
}
\description{
Computes feature specificity scores for each cluster.
}
