% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{correctBatchEffectFastMNN}
\alias{correctBatchEffectFastMNN}
\title{Perform batch correction on AnnData or `SingleCellExperiment`-like inputs.}
\usage{
correctBatchEffectFastMNN(
  adata = NULL,
  batch_attr = NULL,
  layer = NULL,
  reduced_dim = 50,
  MNN_k = 20,
  reduction_out = "MNN",
  BPPARAM = SerialParam(),
  assay_name = NULL,
  ace = NULL
)
}
\description{
Perform batch correction on AnnData or `SingleCellExperiment`-like inputs.
}
