% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decomp.ACTION}
\alias{decomp.ACTION}
\title{Run ACTION_decomposition}
\usage{
decomp.ACTION(
  X,
  k_min,
  k_max,
  specificity_th = -3,
  min_cells_per_arch = 2,
  max_it = 100,
  tol = 1e-16,
  thread_no = 0,
  norm = 1
)
}
\description{
Run ACTION_decomposition
}
