% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal_main.R
\name{.run.mergeArchetypes}
\alias{.run.mergeArchetypes}
\title{Identiy equivalent classes of archetypes and group them together}
\usage{
.run.mergeArchetypes(
  ace,
  reduction_slot = "ACTION",
  C_stacked_slot = "C_stacked",
  H_stacked_slot = "H_stacked",
  normalization = 0,
  merged_suffix = "merged",
  footprint_slot_name = "assigned_archetype",
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Identiy equivalent classes of archetypes and group them together
}
