% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ace.R
\name{filterActionetByAttr}
\alias{filterActionetByAttr}
\title{Filter rows and columns of an ACTIONet-compatible container by observation attribute.}
\usage{
filterActionetByAttr(
  adata = NULL,
  by,
  layer = NULL,
  min_cells_per_feat = NULL,
  min_feats_per_cell = NULL,
  min_umis_per_cell = NULL,
  max_umis_per_cell = NULL,
  assay_name = NULL,
  ace = NULL
)
}
\arguments{
\item{adata}{AnnData or compatible container.}

\item{by}{Observation attribute used to define groups before filtering.}

\item{layer}{Layer to filter on. `NULL` uses `adata$X`.}

\item{assay_name}{Deprecated alias for `layer`.}

\item{ace}{Deprecated alias for `adata`.}
}
\description{
Filter rows and columns of an ACTIONet-compatible container by observation attribute.
}
