% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute.genes.using.archetypes}
\alias{impute.genes.using.archetypes}
\title{Imputing expression of genes by interpolating over archetype profile}
\usage{
impute.genes.using.archetypes(
  adata = NULL,
  genes,
  features_use = NULL,
  ace = NULL
)
}
\arguments{
\item{adata}{AnnData or compatible ACTIONet output container.}

\item{genes}{List of genes to impute}

\item{features_use}{A vector of features of length `n_vars(adata)` or the name of a column of `adata$var` containing the genes given in `genes`.}
}
\value{
A matrix of imputed expression values
}
\description{
Imputing expression of genes by interpolating over archetype profile
}
\examples{
expression_imputed <- impute.genes.using.archetypes(adata, genes)
}
