% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputeFeatures}
\alias{imputeFeatures}
\title{Gene expression imputation using network diffusion.}
\usage{
imputeFeatures(
  adata = NULL,
  features,
  algorithm = c("actionet", "pca"),
  features_use = NULL,
  alpha = 0.85,
  norm_method = "pagerank_sym",
  thread_no = 0,
  max_it = 5,
  layer = "logcounts",
  reduction_slot = "action",
  net_slot = "actionet",
  assay_name = NULL,
  ace = NULL
)
}
\arguments{
\item{adata}{AnnData or compatible container containing output of `runACTIONet()`.}

\item{features}{The list of features to perform imputation on.}

\item{features_use}{A vector of features of length `n_vars(adata)` or the name of a column of `adata$var` containing the features given in `features`.}

\item{alpha}{Depth of diffusion between (0, 1).
The larger it is, the deeper the diffusion, which results in less nonzeros (default = 0.85).}

\item{thread_no}{Number of parallel threads}

\item{layer}{Layer in the container with normalized counts. `NULL` uses `adata$X`.}

\item{diffusion_it}{Number of diffusion iterations (default = 5)}
}
\value{
Imputed feature expression matrix. Column names are set with imputed feature names and rows are observations.
}
\description{
Gene expression imputation using network diffusion.
}
\examples{
adata <- runACTIONet(adata)
imputed.features <- imputeFeatures(adata, c("CD14", "CD19", "CD3G"))
}
