% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{layoutNetwork}
\alias{layoutNetwork}
\title{Performs stochastic force-directed layout on the input graph (ACTIONet)}
\usage{
layoutNetwork(
  G,
  initial_position,
  method = "umap",
  min_dist = 1,
  spread = 1,
  gamma = 1,
  n_epochs = 500L,
  learning_rate = 1,
  seed = 0L,
  thread_no = 0L
)
}
\arguments{
\item{G}{Adjacency matrix of the ACTIONet graph}

\item{n_epochs}{Number of epochs for SGD algorithm (default=100).}

\item{thread_no}{Number of threads (default = 0).}

\item{S_r}{Reduced kernel matrix (is used for reproducible initialization).}

\item{compactness_level}{A value between 0-100, indicating the compactness
of ACTIONet layout (default=50)}
}
\value{
A named list \itemize{
\item coordinates 2D coordinates of vertices.
\item coordinates_3D 3D coordinates of vertices.
\item colors De novo color of nodes inferred from their 3D embedding.
}
}
\description{
Performs stochastic force-directed layout on the input graph (ACTIONet)
}
\examples{
G = buildNetwork(prune.out$H_stacked)
vis.out = layoutNetwrok(G, S_r)
}
