% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_human_mouse_biomart_240924}
\alias{mito_human_mouse_biomart_240924}
\title{Table of mitochodrial transcript features}
\description{
Source data for 'actionet::int_feats_mito'

ENSEMBL-annotated mitochiondrial transcripts from human (hsapiens)
and mouse (mmusculus) transcriptomes. Intended for internal package
use in '.get_mito_feats()'.
}
