% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_SPA}
\alias{run_SPA}
\title{Run successive projections algorithm (SPA) to solve separable NMF}
\usage{
run_SPA(A, k)
}
\arguments{
\item{A}{Input matrix.}

\item{k}{Number of candidate vertices to solve for.}
}
\value{
A named list with entries 'selected_cols' and 'norms'
}
\description{
Run successive projections algorithm (SPA) to solve separable NMF
}
\examples{
H = run_SPA(S_r, 10)
}
