% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{select.top.k.features}
\alias{select.top.k.features}
\title{Report the top-rank features from a given enrichment table}
\usage{
\method{select}{top.k.features}(
  feat_scores,
  top_features = 3,
  normalize = FALSE,
  reorder_columns = TRUE
)
}
\arguments{
\item{top_features}{Number of features to return}

\item{reorder_columns}{Whether to optimally re-order columns of the enrichment table}
}
\value{
Sorted table with the selected top-ranked
}
\description{
Report the top-rank features from a given enrichment table
}
\examples{
feat_scores <- as.matrix(rowMaps(ace)[["arch_feat_spec"]])
enrichment.table.top <- select.top.k.features(feat_scores, 3)
}
