% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotateCells}
\alias{annotateCells}
\title{Infer cell annotations from imputed gene expression for all cells.}
\usage{
annotateCells(
  ace,
  markers,
  method = c("vision", "actionet"),
  features_use = NULL,
  assay_name = "logcounts",
  net_slot = "actionet",
  norm_method = c("pagerank", "pagerank_sym"),
  alpha = 0.85,
  max_it = 5,
  approx = TRUE,
  ignore_baseline = FALSE,
  use_enrichment = TRUE,
  use_lpa = FALSE,
  thread_no = 0
)
}
\arguments{
\item{ace}{ACTIONetExperiment object}

\item{markers}{A named list of marker genes.}

\item{features_use}{A vector of features of length NROW(ace) or the name of a column of rowData(ace) containing the genes given in 'markers'.}

\item{assay_name}{Name of assay for which to impute gene expression (default="logcounts").}

\item{net_slot}{Name of slot in colNets(ace) containing the network to use for gene expression imputation (default="ACTIONet").}

\item{thread_no}{Number of parallel threads used for gene imputation.}

\item{alpha_val}{Random-walk parameter for gene imputation.}
}
\value{
A named list: \itemize{
\item Label: Inferred cell type labels
\item Confidence: Confidence of inferred labels
\item Enrichment: Cell type score matrix.
}
}
\description{
Infer cell annotations from imputed gene expression for all cells.
}
\examples{
data("curatedMarkers_human") # pre-packaged in ACTIONet
markers <- curatedMarkers_human$Blood$PBMC$Monaco2019.12celltypes$marker.genes
annots <- annotate.cells.using.markers(ace, markers = markers)
plot.ACTIONet(ace, annots$Label, annots$Confidence)
}
