% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_decompACTION}
\alias{C_decompACTION}
\title{Run ACTION decomposition algorithm}
\usage{
C_decompACTION(
  S_r,
  k_min = 2L,
  k_max = 30L,
  max_it = 100L,
  tol = 1e-16,
  thread_no = 0L
)
}
\arguments{
\item{S_r}{Input matrix. Usually a reduced representation of the raw data.}

\item{k_min}{Minimum number of archetypes (>= 2) to search for, and the beginning of the search range.}

\item{k_max}{Maximum number of archetypes (<= <b>S_r.n_cols</b>) to search for, and the end of the search range.}

\item{max_it}{Maximum number of iterations for <code>runAA()</code>.}

\item{tol}{Convergence tolerance for <code>runAA()</code>.}

\item{thread_no}{Number of CPU threads to use. If 0, number is automatically determined.}

\item{normalization}{Normalization method to apply on <b>S_r</b> before running ACTION.}
}
\value{
A named list with entries 'C' and 'H', each a list for different values of k
}
\description{
Run ACTION decomposition algorithm
}
\examples{
ACTION.out = runACTION(S_r, k_max = 10)
H8 = ACTION.out$H[[8]]
cell.assignments = apply(H8, 2, which.max)
}
