% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_reduceKernelSparse}
\alias{C_reduceKernelSparse}
\title{Compute reduced kernel matrix}
\usage{
C_reduceKernelSparse(
  S,
  k = 50L,
  svd_alg = 0L,
  max_it = 0L,
  seed = 0L,
  verbose = TRUE
)
}
\arguments{
\item{S}{Input matrix (<em>vars</em> x <em>obs</em>).
May be <code>arma::mat</code> or <code>arma::sp_mat</code>.}

\item{svd_alg}{Singular value decomposition algorithm. See to <code>runSVD()</code> for options.}

\item{max_it}{Maximum number of iterations. Passed to <code>runSVD()</code>.}

\item{seed}{Random seed.}

\item{verbose}{Print status messages.}

\item{dim}{Number of singular vectors to estimate. Passed to <code>runSVD()</code>.}
}
\value{
Field with 5 elements:
- 0: <code>arma::mat</code> Reduced kernel matrix.
- 1: <code>arma::vec</code> Singular values.
- 2: <code>arma::mat</code> Left singular vectors.
- 3: <code>arma::mat</code> <b>A</b> perturbation matrix.
- 4: <code>arma::mat</code> <b>B</b> perturbation matrix.
}
\description{
Compute reduced kernel matrix
}
\examples{
S = logcounts(sce)
reduction.out = reduce(S, reduced_dim = 50)
S_r = reduction.out$S_r
}
