% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_runSVDSparse}
\alias{C_runSVDSparse}
\title{Computes SVD decomposition}
\usage{
C_runSVDSparse(
  A,
  k = 30L,
  max_it = 0L,
  seed = 0L,
  algorithm = 0L,
  verbose = TRUE
)
}
\arguments{
\item{A}{Input matrix ("sparseMatrix")}

\item{k}{Dimension of SVD decomposition}

\item{max_it}{Number of iterations (default=5)}

\item{seed}{Random seed (default=0)}

\item{algorithm}{SVD algorithm to use. Currently supported methods are blah blah blah}
}
\value{
A named list with U, sigma, and V components
}
\description{
This is direct implementation of the randomized SVD algorithm:
From: IRLBA R Package
}
\examples{
A = randn(100, 20)
svd.out = runSVD(A, dim = 3)
U = svd.out$u
}
