% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_specificity.R
\name{computeFeatureSpecificity}
\alias{computeFeatureSpecificity}
\title{Computes feature specificity scores for each cluster.}
\usage{
computeFeatureSpecificity(
  obj,
  labels,
  assay_name = "logcounts",
  map_out_prefix = "cluster",
  return_lower = FALSE,
  thread_no = 0,
  return_raw = FALSE
)
}
\description{
Computes feature specificity scores for each cluster.
}
