% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{correctBatchEffectFastMNN}
\alias{correctBatchEffectFastMNN}
\title{Perform batch correction on `ACTIONetExperiment` and `SingleCellExperiment` objects.}
\usage{
correctBatchEffectFastMNN(
  ace,
  batch_attr = NULL,
  assay_name = NULL,
  reduced_dim = 50,
  MNN_k = 20,
  reduction_out = "MNN",
  BPPARAM = SerialParam()
)
}
\description{
Perform batch correction on `ACTIONetExperiment` and `SingleCellExperiment` objects.
}
