% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{reduce.ace}
\alias{reduce.ace}
\title{Takes a `ACTIONetExperiment` object and adds the reduced kernel matrix}
\usage{
reduce.ace(
  ace,
  reduced_dim = 50,
  assay_name = "logcounts",
  reduction_out = "action",
  algorithm = c("irlb", "halko", "feng"),
  max_it = NULL,
  seed = 0,
  verbose = TRUE
)
}
\arguments{
\item{ace}{ACTIONetExperiment object.}

\item{reduced_dim}{Dimension of reduced kernel matrix. Number of singular vectors to estimate. Passed to <code>runSVD()</code>.}

\item{assay_name}{Name of assay to reduce.}

\item{reduction_out}{Name of slot to store reduced output.}

\item{algorithm}{Singular value decomposition algorithm. Passed to <code>runSVD()</code>.}

\item{max_it}{Number of SVD iterations. If `NULL`: 1000 for "ilrb", 5 otherwise.}

\item{seed}{Random seed.}
}
\value{
ACTIONetExperiment object with reduction in colMaps(ace).
}
\description{
Takes a `ACTIONetExperiment` object and adds the reduced kernel matrix
}
\examples{
ace <- import.ace.from.10X(input_path)
ace <- reduce.ace(ace)
}
