% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_action.R
\name{reduceKernel}
\alias{reduceKernel}
\title{Compute the reduced kernel matrix and decomposition}
\usage{
reduceKernel(
  obj,
  k = 50,
  algorithm = c("irlb", "halko", "feng"),
  max_it = NULL,
  seed = 0,
  verbose = TRUE,
  assay_name = "logcounts",
  reduction_slot = "action",
  return_raw = FALSE
)
}
\arguments{
\item{obj}{ACTIONetExperiment, SummarizedExperiment, SingleCellExperiment or matrix.}

\item{k}{Dimension of reduced kernel matrix. Number of singular vectors to estimate. Passed to <code>runSVD()</code>.}

\item{algorithm}{Singular value decomposition algorithm. Passed to <code>runSVD()</code>.}

\item{max_it}{Number of SVD iterations. If `NULL`: 1000 for "ilrb", 5 otherwise.}

\item{seed}{Random seed.}

\item{verbose}{Print status messages.}

\item{assay_name}{Name of assay to reduce.}

\item{reduction_slot}{Slot of colMaps(ace) in which to stored reduced kernel matrix. Also the prefix of slots to in which to store output components of the decomposition. (ACTIONetExperiment output only).}

\item{return_raw}{Return raw output regardless of 'obj' type.}
}
\value{
ACTIONetExperiment object with reduction in colMaps(ace).
}
\description{
Compute the reduced kernel matrix and decomposition
}
