% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runACTIONet}
\alias{runACTIONet}
\title{Run main ACTIONet pipeline}
\usage{
runACTIONet(
  ace,
  k_min = 2,
  k_max = 30,
  assay_name = "logcounts",
  reduction_slot = "action",
  net_slot_out = "actionet",
  min_obs = 2,
  max_it = 50,
  spec_th = -3,
  network_metric = "jsd",
  network_algorithm = "k*nn",
  network_density = 1,
  network_ef_construction = 200,
  network_ef = 200,
  mutual_edges_only = TRUE,
  layout_method = c("umap", "tumap", "largevis"),
  layout_epochs = 100,
  layout_spread = 1,
  layout_min_dist = 1,
  layout_3d = TRUE,
  layout_parallel = TRUE,
  compute_specificity_parallel = FALSE,
  thread_no = 0,
  seed = 0
)
}
\arguments{
\item{ace}{`ACTIONetExperiment` containing an appropriate reduction in 'reduction_slot'.}

\item{k_min}{Minimum depth of decompositions. (default=2)}

\item{k_max}{Maximum depth of decompositions. (default=30)}

\item{assay_name}{Name of assay to be used. (default='logcounts')}

\item{reduction_slot}{Slot in colMaps(ace) containing reduced kernel. (default='ACTION')}

\item{net_slot_out}{Name of slot in colMaps(ace) to store ACTIONet adjacency matrix. (default='ACTIONet')}

\item{min_obs}{Minimum number of observations required to construct an archetype. (default=2)}

\item{max_it}{Maximum number of iterations for ACTION algorithm. (default=50)}

\item{spec_th}{Defines the stringency of pruning nonspecific archetypes.
The larger the value, the more archetypes will be filtered out. (default=-3)}

\item{network_metric}{Distance metric with which to compute cell-to-cell similarity during network construction. Options are 'jsd' (Jensen-Shannon divergence), L2-norm ('l2'), and inner product ('ip'). (default='jsd')}

\item{network_algorithm}{Algorithm to use for network construction. Options are k-nearest neighbors ('knn') and k*-nearest neighbors ('k*nn'). (default='k*nn')}

\item{network_density}{Density factor of ACTIONet graph. (default=1)}

\item{network_ef_construction}{HNSW construction search breadth for network construction. For `network_algorithm="k*nn"`, the effective value is `max(network_ef_construction, kNN)`. (default=200)}

\item{network_ef}{HNSW query search breadth for network construction. For `network_algorithm="k*nn"`, the effective value is `max(network_ef, kNN)`. (default=200)}

\item{mutual_edges_only}{Whether to enforce edges to be mutually-nearest-neighbors. (default=TRUE)}

\item{layout_method}{Algorithm for computing plot layout. Options are UMAP ("umap") or t-UMAP ("tumap"). (default="umap")}

\item{layout_epochs}{Number of epochs for SGD algorithm. (default=250)}

\item{layout_parallel}{Run layout construction using multiple cores. May result in marginally different outputs across runs due to parallelization-induced randomization. (default=TRUE)}

\item{compute_specificity_parallel}{Run feature specificity enrichment using multiple cores. Setting this to `TRUE` on large datasets may cause an out of memory crash. (default=FALSE)}

\item{thread_no}{Number of parallel threads. (default=0)}

\item{seed}{Seed for random initialization. (default=0)}
}
\value{
An ACTIONetExperiment object
}
\description{
Run main ACTIONet pipeline
}
\examples{
ace <- runACTIONet(ace)
}
