% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_anndata_adapter.R
\name{toACTIONetExperiment}
\alias{toACTIONetExperiment}
\title{Convert AnnData to `ACTIONetExperiment`.}
\usage{
toACTIONetExperiment(adata)
}
\arguments{
\item{adata}{AnnData or compatible input.}
}
\value{
An `ACTIONetExperiment` object.
}
\description{
Convert AnnData to `ACTIONetExperiment`.
}
